// apps/web/lib/api/generated/models/createOrgUnitDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * llame API
 * llame auth and domain API
 * OpenAPI spec version: 0.1
 */
import type { CreateOrgUnitDtoType } from "./createOrgUnitDtoType";

export interface CreateOrgUnitDto {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  type?: CreateOrgUnitDtoType;
}